safari-600.7-branch/LayoutTests/canvas/philip/tests/security.dataURI.html

<!DOCTYPE html>
<title>Canvas test: security.dataURI</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
var data = canvas.toDataURL();
ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
var img = new Image();
deferTest();
img.onload = wrapFunction(function ()
{
    ctx.drawImage(img, 0, 0);
    canvas.toDataURL(); // should be permitted
    _assertPixel(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");
});
img.src = data;


});
</script>
