LayoutTests/canvas/philip/tests/security.dataURI-expected.txt

< [index] >
security.dataURI
data: URIs do not count as different-origin, and do not taint the canvas
References: Bugzilla
Actual output:
Expected output:

Passed
